requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
PyJWT
PyPDF2
faiss-cpu
python-multipart